+++ src/core/cache/decorators/persistent/engines/interface.ts
@@ -16,10 +16,15 @@ export interface AbstractPersistentEngine<V = unknown> {
 	 * Initializes a new cache instance from the past one
 	 * @param cache
 	 */
-	initCache?(cache: Cache<V>): CanPromise<void>;
+	initCache?(cache: Cache<unknown, V>): CanPromise<void>;
 }
 
 export abstract class AbstractPersistentEngine<V = unknown> {
+	/**
+	 * Index with keys and TTL-s of stored values
+	 */
+	ttlIndex: Dictionary<number> = Object.createDict();
+
 	/**
 	 * API for async operations
 	 */